+++ src/main/java/juja/microservices/gamification/slackbot/model/achievements/DailyAchievement.java
@@ -1,4 +1,4 @@
-package juja.microservices.gamification.slackbot.model;
+package juja.microservices.gamification.slackbot.model.achievements;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
