+++ src/main/java/juja/microservices/gamification/slackbot/model/achievements/InterviewAchievement.java
@@ -1,0 +1,25 @@
+package juja.microservices.gamification.slackbot.model.achievements;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.ToString;
+
+/**
+ * @author Artem
+ */
+
+@Getter
+@ToString
+public class InterviewAchievement {
+    @JsonProperty
+    private String from;
+    @JsonProperty
+    private String description;
+
+    public InterviewAchievement(String from, String description) {
+        this.from = from;
+        this.description = description;
+    }
+}
+
+
