+++ src/main/java/juja/microservices/gamification/slackbot/model/achievements/ThanksAchievement.java
@@ -1,4 +1,4 @@
-package juja.microservices.gamification.slackbot.model;
+package juja.microservices.gamification.slackbot.model.achievements;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import juja.microservices.gamification.slackbot.exceptions.WrongCommandFormatException;
